[Refactor] align Fly dialect modules with upstream naming (#305)
---------
