+++ python/flydsl/compiler/backends/rocm.py
@@ -71,8 +71,7 @@ class RocmBackend(BaseBackend):
 
     def native_lib_patterns(self) -> List[str]:
         return [
-            "_fly*.so",
-            "_fly_rocdl*.so",
+            "_mlirDialectsFly*.so",
             "libFly*.so",
             "libfly_jit_runtime.so",
             "libmlir_rocm_runtime.so",
