+++ python/flydsl/expr/rocdl.py
@@ -17,7 +17,7 @@ Example:
     >>> rocdl.barrier()
 """
 
-from .._mlir._mlir_libs._fly_rocdl import CopyOpCDNA3BufferCopyType, MmaAtomCDNA3_MFMAType
+from .._mlir._mlir_libs._mlirDialectsFlyROCDL import CopyOpCDNA3BufferCopyType, MmaAtomCDNA3_MFMAType
 from .._mlir.dialects.rocdl import *  # noqa: F401,F403
 from .._mlir.extras import types as T
