+++ python/mlir_flydsl/dialects/fly.py
@@ -6,4 +6,4 @@ from ._fly_ops_gen import *
 from ._fly_ops_gen import _Dialect
 from ._fly_enum_gen import *
 
-from .._mlir_libs._fly import *
+from .._mlir_libs._mlirDialectsFly import *
