+++ python/mlir_flydsl/dialects/fly_rocdl.py
@@ -6,4 +6,4 @@ from ._fly_rocdl_ops_gen import *
 from ._fly_rocdl_ops_gen import _Dialect
 from ._fly_rocdl_enum_gen import *
 
-from .._mlir_libs._fly_rocdl import *
+from .._mlir_libs._mlirDialectsFlyROCDL import *
